default

+++ src/main/java/org/example/model/Order.java
@@ -14,7 +14,7 @@ public class Order implements Serializable {
 
   private static final long serialVersionUID = 62345888887777L;
 
-  private Integer id,petId,quantity;
+  private Long id,petId,quantity;
   private String shipDate,status;
   private Boolean complete;
 }
