+++ src/main/java/org/example/model/User.java
@@ -14,6 +14,6 @@ public class User implements Serializable {
 
   private static final long serialVersionUID = 239475831256347L;
 
-  private Integer id,userStatus;
+  private Long id,userStatus;
   private String username,firstName,lastName,email,password,phone;
 }
